// services/resumes/lib/generated.graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  _FieldSet: any;
};

export type AdditionalEntityFields = {
  path?: InputMaybe<Scalars['String']>;
  type?: InputMaybe<Scalars['String']>;
};

export type Application = {
  __typename?: 'Application';
  /** Filename of the user's resume, likely as a  */
  resumeFilename?: Maybe<Scalars['String']>;
  /** Signed read URL for the user's resume */
  resumeUrl?: Maybe<Scalars['String']>;
};

export type Mutation = {
  __typename?: 'Mutation';
  /**
   * Retrieves a URL where a file may be uploaded. Individual users are only permitted
   * to upload one file, so as a side effect, requesting a new file upload URL will delete
   * any file owned by the user in storage. The corresponding file will be named after the
   * user'd primary ID, followed by the file extension provided.
   */
  fileUploadUrl?: Maybe<Scalars['String']>;
  /**
   * Retrieves a signed read url for a zip file of resumes matching all of the provided filter criteria.
   *
   * This is implemented as a mutation, as it is expensive to call.
   */
  resumeDumpUrl?: Maybe<Scalars['String']>;
};


export type MutationFileUploadUrlArgs = {
  contentType: Scalars['String'];
  fileExtension: Scalars['String'];
};


export type MutationResumeDumpUrlArgs = {
  filterCriteria?: InputMaybe<UserFilterCriteria>;
};
// ...
export type ReferenceResolver<TResult, TReference, TContext> = (
      reference: TReference,
      context: TContext,
      info: GraphQLResolveInfo
    ) => Promise<TResult> | TResult;

      type ScalarCheck<T, S> = S extends true ? T : NullableCheck<T, S>;
      type NullableCheck<T, S> = Maybe<T> extends T ? Maybe<ListCheck<NonNullable<T>, S>> : ListCheck<T, S>;
      type ListCheck<T, S> = T extends (infer U)[] ? NullableCheck<U, S>[] : GraphQLRecursivePick<T, S>;
      export type GraphQLRecursivePick<T, S> = { [K in keyof T & keyof S]: ScalarCheck<T[K], S[K]> };
    

export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;

export type NextResolverFn<T> = () => Promise<T>;

export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (
  next: NextResolverFn<TResult>,
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => TResult | Promise<TResult>;

/** Mapping between all available schema types and the resolvers types */
export type ResolversTypes = {
  AdditionalEntityFields: AdditionalEntityFields;
  String: ResolverTypeWrapper<Scalars['String']>;
  Application: ResolverTypeWrapper<Application>;
  Mutation: ResolverTypeWrapper<{}>;
  UserFilterCriteria: UserFilterCriteria;
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  AdditionalEntityFields: AdditionalEntityFields;
  String: Scalars['String'];
  Application: Application;
  Mutation: {};
  UserFilterCriteria: UserFilterCriteria;
  Boolean: Scalars['Boolean'];
};
// ...
export type ApplicationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Application'] = ResolversParentTypes['Application']> = {
  __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Application']>, { __typename: 'Application' } & GraphQLRecursivePick<ParentType, {"id":true}>, ContextType>;
  resumeFilename?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  resumeUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  fileUploadUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, RequireFields<MutationFileUploadUrlArgs, 'contentType' | 'fileExtension'>>;
  resumeDumpUrl?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType, Partial<MutationResumeDumpUrlArgs>>;
};

export type Resolvers<ContextType = any> = {
  Application?: ApplicationResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
};